requirements.txt
streamlit
pypdf
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-google-genai
sentence-transformers
faiss-cpu



#Create a virtual environment and install dependencies
# python -m venv venv
# source venv/bin/activate  # On Windows use `venv\Scripts\activate`  
#pip install python-dotenv  
#How to run the app
#streamlit run chatbot.py
#pip install -r requirements.txt
# python -m pip install python-dotenv
# python chatbot.py
